// src/modules/evaluate/evaluate.dto.ts
export class DTO_RQ_Evaluate {
    desc: string;
    rating: number;
    company_id: number;
    trip_id: number;
    ticket_id: number;
    account_id: string;
    account_name: string;
    ticket_phone: string;
    account_email: string;
    account_avatar: string;
}
export class DTO_RP_Evaluate {
    id: number;
    desc: string;
    rating: number;
    company_id: number;
    trip_id: number;
    ticket_id: number;
    account_id: string;
    account_name: string;
    ticket_phone: string;
    account_email: string;
    account_avatar: string;
    created_at: string;
}